rename StartOptions to GlobalConfig
