+++ ui/packages/tidb-dashboard-for-clinic-cloud/public/index.html
@@ -128,7 +128,9 @@
       startDashboard({
         clientOptions: {
           apiPathBase,
-          apiToken,
+          apiToken
+        },
+        clusterInfo: {
           orgId,
           clusterId
         }
